feat: Add --purge option to remove cache directory
Add a new --purge command-line option to remove the cache directory
at $HOME/.cache/cargo-unmaintained/v2. This enables users to easily
clean up the on-disk cache without having to manually find and delete
the directory.

The option is only available on non-Windows platforms with the
on-disk-cache feature enabled, since Windows uses temporary caching.